host and non shared in kosambi

+++ opencv/host/affine.cu
@@ -5,10 +5,10 @@ using namespace cv;
 class AffineTransformFunctor
 {
 public:
-  thrust::host_block_2d<uchar> *transformMatrix;
+  thrust::host_block_2d<float> *transformMatrix;
   thrust::host_block_2d<uchar> *outBlock;
 
-  AffineTransformFunctor(thrust::host_block_2d<uchar> * tm,thrust::host_block_2d<uchar> * outBlock)
+  AffineTransformFunctor(thrust::host_block_2d<float> * tm,thrust::host_block_2d<uchar> * outBlock)
   {
     this->transformMatrix = tm;
     this->outBlock = outBlock;
@@ -18,7 +18,6 @@ public:
     int x_out, y_out;
     x_out = (int)((*transformMatrix)[0][0]*inputWindow.start_x+(*transformMatrix)[0][1]*inputWindow.start_y+(*transformMatrix)[0][2]*1);
     y_out = (int)((*transformMatrix)[1][0]*inputWindow.start_x+(*transformMatrix)[1][1]*inputWindow.start_y+(*transformMatrix)[1][2]*1);
-
     (*outBlock)[y_out][x_out]=inputWindow[0][0];
   }
 };
@@ -48,12 +47,12 @@ int main(int argc, char const *argv[]) {
   warp_mat = getAffineTransform( srcTri, dstTri );
   warp_mat.convertTo(warp_mat,CV_32FC1);
   //Move Warp Matrix to Device
-  thrust::host_block_2d<uchar> warp_block(warp_mat.cols,warp_mat.rows);
+  thrust::host_block_2d<float> warp_block(warp_mat.cols,warp_mat.rows);
   for(int i = 0; i< warp_mat.rows;i++)
   {
     for(int j = 0; j<warp_mat.cols;j++)
     {
-      warp_block[i][j]=warp_mat.at<uchar>(i,j);
+      warp_block[i][j]=warp_mat.at<float>(i,j);
     }
   }
   //Create Windows For Indexing
